fix: adding pricing service and swap quote endpoint
