+++ modules/l0/src/main/scala/org/amm_metagraph/l0/rewards/package.scala
@@ -1,7 +1,6 @@
 package org.amm_metagraph.l0
 
 import eu.timepit.refined.api.Refined
-import eu.timepit.refined.generic.Equal
 import eu.timepit.refined.numeric.Interval
 
 package object rewards {
