+++ modules/shared_data/src/main/scala/org/amm_metagraph/shared_data/types/LiquidityPool.scala
@@ -17,7 +17,6 @@ import derevo.derive
 import eu.timepit.refined.types.numeric.PosLong
 import io.circe.refined._
 import io.estatico.newtype.macros.newtype
-import org.amm_metagraph.shared_data.refined._
 import org.amm_metagraph.shared_data.types.DataUpdates.LiquidityPoolUpdate
 import org.amm_metagraph.shared_data.types.States._
 
@@ -99,20 +98,4 @@ object LiquidityPool {
     getLiquidityPoolCalculatedState(state).pending.collect {
       case pendingSpend: PendingSpendAction[LiquidityPoolUpdate] => pendingSpend
     }
-
-  def getLiquidityPoolPrices(liquidityPool: LiquidityPool): (Long, Long) = {
-    val priceTokenA = (BigDecimal(liquidityPool.tokenB.amount.value) /
-      BigDecimal(liquidityPool.tokenA.amount.value))
-      .setScale(0, BigDecimal.RoundingMode.HALF_UP)
-      .toLong
-      .toTokenAmountFormat
-
-    val priceTokenB = (BigDecimal(liquidityPool.tokenA.amount.value) /
-      BigDecimal(liquidityPool.tokenB.amount.value))
-      .setScale(0, BigDecimal.RoundingMode.HALF_UP)
-      .toLong
-      .toTokenAmountFormat
-
-    (priceTokenA, priceTokenB)
-  }
 }
